+++ tfjs-core/src/public/chained_ops/register_all_chained_ops.ts
@@ -23,6 +23,7 @@ import './concat';
 import './conv1d';
 import './conv2d';
 import './conv2d_transpose';
+import './cumsum';
 import './depth_to_space';
 import './depthwise_conv2d';
 import './depthwise_conv2D_deprecated';
