+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -31,6 +31,7 @@ const CHAINED_OPS = [
   'concat',
   'conv1d',
   'conv2d',
+  'cumsum',
   'conv2dTranspose',
   'depthToSpace',
   'depthwiseConv2d',
